requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.18
tabulate==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
